documentai/v1/python/google-cloud-documentai/requirements.txt
google-cloud-documentai==3.11.0
